Change behaviour to allow user to change operator
Previously clicking a second operator would force an implicit
calculation to be made using the prefilled display value of 0, so that
"5, x, + 3" would return ((5*0)+3). It now has a more intuitive behaviour
of successive operators replacing the previous if no num2 was inputted.
This gives "5, x, + 3" as (5+3) instead.

Note that both still behave the same if passed a series of inputs such as
"5, x, 2, +, 3" as both behaviours will handle it operation by operation
as it is inputted: ((5*2)+3), as was detailed in the spec